+++ lib/types/helpers/collection-helpers.ts
@@ -1,9 +1,15 @@
 /**
+ * Extracts the element type from an array type.
+ * @template A - The array type.
+ * @returns The element type of the array.
+ */
+export type ArrayElement<A> = A extends Array<infer Element> ? Element : never;
+
+/**
  * Represents a type that can be either a single {@link Value} or an {@link Array} of {@link Value}.
  *
  * @template Value - The type of the value(s) in the array.
  */
-
 export type ArrayLike<Value> = Value | Value[];
 /**
  * Represents a type that can either be a single {@link Value} or a {@link Set} of {@link Value}.
@@ -36,7 +42,31 @@ export type CollectionLike<Value, Key = unknown> =
  * Represents a non-empty {@link Array} type.
  * @template T - The type of elements in the array.
  */
-export type NonEmptyArray<T> = [T, ...T[]];
+export type NonEmptyArray<T> = ArrayWithAtLeast1<T>;
+
+/**
+ * Represents an array type with at least one element.
+ * @template T - The type of the array elements.
+ */
+export type ArrayWithAtLeast1<T> = [T, ...T[]];
+
+/**
+ * Represents an array with at least two elements of type T.
+ * @template T The type of elements in the array.
+ */
+export type ArrayWithAtLeast2<T> = [T, ...ArrayWithAtLeast1<T>];
+
+/**
+ * Represents an array with at least three elements of type T.
+ * @template T The type of elements in the array.
+ */
+export type ArrayWithAtLeast3<T> = [T, ...ArrayWithAtLeast2<T>];
+
+/**
+ * Represents an array with at least four elements of type T.
+ * @template T The type of elements in the array.
+ */
+export type ArrayWithAtLeast4<T> = [T, ...ArrayWithAtLeast3<T>];
 
 /**
  * Represents a recursive {@link Array} that can contain values of type {@link T}, undefined, or other recursive arrays.